Fix state restoration when going backward
Restoring the state of the selected values in the sidebar
expected there to be values in targetResources. If this is
a backwards navigation, there are no targetResources,
and so restoring state from by starting from startValue and
sprinkling in stuff from targetResources will fail. Need to sprinkle
in stuff from value(), hence value() needs to be a real obj, not
a resource id.